application/src/main/java/com/juliomesquita/admin/catalog/application/usecases/category/delete/DefaultDeleteCategoryUseCase.java
package com.juliomesquita.admin.catalog.application.usecases.category.delete;

import com.juliomesquita.admin.catalog.domain.category.CategoryGateway;
import com.juliomesquita.admin.catalog.domain.category.CategoryId;

import java.util.Objects;

public class DefaultDeleteCategoryUseCase extends DeleteCategoryUseCase {
    private final CategoryGateway categoryGateway;

    public DefaultDeleteCategoryUseCase(final CategoryGateway categoryGateway) {
        this.categoryGateway = Objects.requireNonNull(categoryGateway);
    }

    @Override
    public void execute(final String aCommand) {
        this.categoryGateway.deleteById(CategoryId.from(aCommand));
    }
}
